requirements.txt
Django==3.2.5
django-embed-video==1.3.1
django-extensions==2.2.8
django-simplecms==1.0.5
easy-thumbnails==2.7
idna==2.9
libsass==0.19.4
Markdown==3.2.1
mollie-api-python==2.3.0
psycopg2==2.9.5
pylibmc==1.6.3
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
